pyproject.toml
[project]
name = "shadowbox-generator"
version = "0.1.0"
description = "Transform TCG card illustrations into interactive 3D shadowbox displays using AI-powered depth estimation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "2ROSSO" }]
keywords = ["tcg", "shadowbox", "depth-estimation", "3d", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]

dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "vedo>=2024.5.0",
    "opencv-python>=4.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "ipympl>=0.9.0",
    "plotly>=5.18.0",  # Jupyter 3D rendering
]
gui = [
    "PyQt6>=6.5.0",
]
triposr = [
    # 試験搭載 (Experimental): 既存依存と競合する場合があります。
    # 別の仮想環境での利用を推奨します。
    # TripoSR本体は別途手動インストールが必要:
    #   git clone https://github.com/VAST-AI-Research/TripoSR.git
    #   pip install -r TripoSR/requirements.txt
    "trimesh>=4.0.0",     # Mesh processing
    "rtree>=1.0.0",       # R-tree for trimesh ray casting fallback
    "omegaconf>=2.3.0",   # Configuration
    "einops>=0.7.0",      # Tensor operations
    "pyrender>=0.1.45",   # Offscreen rendering for depth extraction
    "rembg>=2.0.50",      # 背景除去
    "onnxruntime>=1.24.0",  # rembg依存（1.24+ で Python 3.11+ 対応済み）
    "pymatting>=1.1.15",  # rembg依存（1.1.15+ で Python 3.11+ 対応の numba を使用）
    # Note: pyrender requires OpenGL. Falls back to trimesh ray casting if unavailable.
]
all = [
    # triposr は試験搭載のため除外（明示的に --extra triposr で指定可能）
    "shadowbox-generator[jupyter,gui]",
]

[project.urls]
Homepage = "https://github.com/2ROSSO/tcg-shadowbox-generator"
Repository = "https://github.com/2ROSSO/tcg-shadowbox-generator"

[project.scripts]
shadowbox = "shadowbox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shadowbox"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-qt>=4.2.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "vedo.*",
    "transformers.*",
    "sklearn.*",
    "cv2.*",
    "tsr.*",
    "trimesh.*",
    "rembg.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=shadowbox --cov-report=term-missing -m 'not slow and not integration and not gui'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring external resources",
    "gui: marks tests that require a display (PyQt6/pytest-qt)",
]
